requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1

# MetaTrader5 - тільки для Windows
# Для Linux: встановіть окремо на Windows машині або використайте requirements-windows.txt
# pip install MetaTrader5==5.0.45  # Тільки на Windows
